Faz 2-3: Security hardening + Decimal para migration

+++ lib/features/comparison/presentation/widgets/comparison_result_card.dart
@@ -134,7 +134,7 @@ class _ComparisonResultCardState extends State<ComparisonResultCard>
                       crossAxisAlignment: CrossAxisAlignment.end,
                       children: [
                         Text(
-                          _tryFormatter.format(calc.finalValueTry),
+                          _tryFormatter.format(calc.finalValueTry.toDouble()),
                           style: Theme.of(context).textTheme.bodyMedium
                               ?.copyWith(fontWeight: FontWeight.bold),
                         ),
